+++ remy/internal/injector/standard_injector.go
@@ -2,19 +2,20 @@ package injector
 
 import (
 	"github.com/wrapped-owls/goremy-di/remy/internal/types"
+	"github.com/wrapped-owls/goremy-di/remy/internal/utils"
 )
 
 type (
 	StdInjector struct {
-		allowOverride     bool
-		generifyInterface bool
-		parentInjector    types.DependencyRetriever
-		bindStorage       types.Storage[types.BindKey]
-		instanceStorage   types.Storage[types.BindKey]
+		allowOverride   bool
+		reflectOpts     types.ReflectionOptions
+		parentInjector  types.DependencyRetriever
+		bindStorage     types.Storage[types.BindKey]
+		instanceStorage types.Storage[types.BindKey]
 	}
 )
 
-func New(canOverride bool, generifyInterface bool, parent ...types.Injector) *StdInjector {
+func New(canOverride bool, reflectOpts types.ReflectionOptions, parent ...types.Injector) *StdInjector {
 	var parentInjector types.Injector
 	if len(parent) > 0 {
 		parentInjector = parent[0]
@@ -22,8 +23,8 @@ func New(canOverride bool, generifyInterface bool, parent ...types.Injector) *St
 	return &StdInjector{
 		allowOverride:   canOverride,
 		parentInjector:  parentInjector,
-		bindStorage:     NewElementsStorage[types.BindKey](canOverride, generifyInterface),
-		instanceStorage: NewElementsStorage[types.BindKey](canOverride, generifyInterface),
+		bindStorage:     NewElementsStorage[types.BindKey](canOverride, reflectOpts),
+		instanceStorage: NewElementsStorage[types.BindKey](canOverride, reflectOpts),
 	}
 }
 
@@ -33,11 +34,11 @@ func (s *StdInjector) SubInjector(overrides ...bool) types.Injector {
 		canOverride = overrides[0]
 	}
 
-	return New(canOverride, s.generifyInterface, s)
+	return New(canOverride, s.reflectOpts, s)
 }
 
-func (s StdInjector) ShouldGenerifyInterface() bool {
-	return s.generifyInterface
+func (s StdInjector) ReflectOpts() types.ReflectionOptions {
+	return s.reflectOpts
 }
 
 func (s *StdInjector) Bind(key types.BindKey, value any) {
@@ -48,20 +49,24 @@ func (s *StdInjector) BindNamed(name string, bType types.BindKey, value any) {
 	s.bindStorage.SetNamed(bType, name, value)
 }
 
-func (s StdInjector) RetrieveBind(key types.BindKey) (any, bool) {
-	result, ok := s.bindStorage.Get(key)
-	if !ok && s.parentInjector != nil {
-		return s.parentInjector.RetrieveBind(key)
+func (s StdInjector) RetrieveBind(key types.BindKey) (result any, err error) {
+	if result, err = s.bindStorage.Get(key); err != nil && s.parentInjector != nil {
+		result, err = s.parentInjector.RetrieveBind(key)
+		if err != nil {
+			err = utils.ErrNoElementFoundInsideOrParent
+		}
 	}
-	return result, ok
+	return
 }
 
-func (s StdInjector) RetrieveNamedBind(name string, bType types.BindKey) (any, bool) {
-	result, ok := s.bindStorage.GetNamed(bType, name)
-	if !ok && s.parentInjector != nil {
-		return s.parentInjector.RetrieveNamedBind(bType, name)
+func (s StdInjector) RetrieveNamedBind(name string, bType types.BindKey) (result any, err error) {
+	if result, err = s.bindStorage.GetNamed(bType, name); err != nil && s.parentInjector != nil {
+		result, err = s.parentInjector.RetrieveNamedBind(bType, name)
+		if err != nil {
+			err = utils.ErrNoElementFoundInsideOrParent
+		}
 	}
-	return result, ok
+	return
 }
 
 func (s *StdInjector) Set(key types.BindKey, value any) {
@@ -72,18 +77,22 @@ func (s *StdInjector) SetNamed(elementType types.BindKey, name string, value any
 	s.instanceStorage.SetNamed(elementType, name, value)
 }
 
-func (s StdInjector) GetNamed(bindKey types.BindKey, name string) (any, bool) {
-	result, ok := s.instanceStorage.GetNamed(bindKey, name)
-	if !ok && s.parentInjector != nil {
-		return s.parentInjector.GetNamed(bindKey, name)
+func (s StdInjector) GetNamed(bindKey types.BindKey, name string) (result any, err error) {
+	if result, err = s.instanceStorage.GetNamed(bindKey, name); err != nil && s.parentInjector != nil {
+		result, err = s.parentInjector.GetNamed(bindKey, name)
+		if err != nil {
+			err = utils.ErrNoElementFoundInsideOrParent
+		}
 	}
-	return result, ok
+	return
 }
 
-func (s StdInjector) Get(bindKey types.BindKey) (any, bool) {
-	result, ok := s.instanceStorage.Get(bindKey)
-	if !ok && s.parentInjector != nil {
-		return s.parentInjector.Get(bindKey)
+func (s StdInjector) Get(bindKey types.BindKey) (result any, err error) {
+	if result, err = s.instanceStorage.Get(bindKey); err != nil && s.parentInjector != nil {
+		result, err = s.parentInjector.Get(bindKey)
+		if err != nil {
+			err = utils.ErrNoElementFoundInsideOrParent
+		}
 	}
-	return result, ok
+	return
 }