+++ remy/internal/injector/standard_injector_test.go
@@ -10,7 +10,7 @@ import (
 
 func TestStdInjector_SubInjector(t *testing.T) {
 	const strFirstHalf = "the counter is at"
-	parent := New(false, false)
+	parent := New(false, types.ReflectionOptions{})
 	subInjector := parent.SubInjector(false)
 
 	var counter uint8 = 0
@@ -20,12 +20,12 @@ func TestStdInjector_SubInjector(t *testing.T) {
 	}))
 
 	Register(subInjector, binds.Factory(func(retriever types.DependencyRetriever) string {
-		return fmt.Sprintf("%s %d", strFirstHalf, Get[uint8](retriever))
+		return fmt.Sprintf("%s %d", strFirstHalf, TryGet[uint8](retriever))
 	}))
 
 	for i := 0; i < 255; i++ {
 		expected := fmt.Sprintf("%s %d", strFirstHalf, i+1)
-		if result := Get[string](subInjector); result != expected {
+		if result := TryGet[string](subInjector); result != expected {
 			t.Errorf("sub-injector is not calling parent injector correctly. Received: `%s`; Expected: `%s`", result, expected)
 			t.FailNow()
 		}
@@ -34,7 +34,7 @@ func TestStdInjector_SubInjector(t *testing.T) {
 
 func TestStdInjector_SubInjector__OverrideParent(t *testing.T) {
 	const strFirstHalf = "The totally value of it is"
-	parent := New(false, false)
+	parent := New(false, types.ReflectionOptions{})
 	subInjector := parent.SubInjector(false)
 
 	Register(parent, binds.Factory(func(retriever types.DependencyRetriever) uint8 {
@@ -42,11 +42,11 @@ func TestStdInjector_SubInjector__OverrideParent(t *testing.T) {
 	}))
 
 	Register(subInjector, binds.Factory(func(retriever types.DependencyRetriever) string {
-		return fmt.Sprintf("%s %d", strFirstHalf, Get[uint8](retriever))
+		return fmt.Sprintf("%s %d", strFirstHalf, TryGet[uint8](retriever))
 	}))
 
 	expected := fmt.Sprintf("%s 101", strFirstHalf)
-	if result := Get[string](subInjector); result != expected {
+	if result := TryGet[string](subInjector); result != expected {
 		t.Errorf("sub-injector is not calling parent injector correctly. Received: `%s`; Expected: `%s`", result, expected)
 		t.FailNow()
 	}
@@ -57,13 +57,13 @@ func TestStdInjector_SubInjector__OverrideParent(t *testing.T) {
 	}))
 
 	expected = fmt.Sprintf("%s 42", strFirstHalf)
-	if result := Get[string](subInjector); result != expected {
+	if result := TryGet[string](subInjector); result != expected {
 		t.Errorf("sub-injector is not calling parent injector correctly. Received: `%s`; Expected: `%s`", result, expected)
 		t.FailNow()
 	}
 
 	// Checks if parent still returns the same old value
-	parentResult := Get[uint8](parent)
+	parentResult := TryGet[uint8](parent)
 	if parentResult != 101 {
 		t.Errorf("parent value was overrided, it should not occur")
 	}