+++ remy/internal/injector/storage.go
@@ -1,28 +1,31 @@
 package injector
 
-import "github.com/wrapped-owls/goremy-di/remy/internal/utils"
+import (
+	"github.com/wrapped-owls/goremy-di/remy/internal/types"
+	"github.com/wrapped-owls/goremy-di/remy/internal/utils"
+)
 
 // ElementsStorage holds all dependencies
 type (
 	genericAnyMap[T comparable]   map[T]any
 	ElementsStorage[T comparable] struct {
 		allowOverride     bool
-		generifyInterface bool
+		generifyInterface types.ReflectionOptions
 		namedElements     map[string]genericAnyMap[T]
 		elements          genericAnyMap[T]
 	}
 )
 
-func NewElementsStorage[T comparable](allowOverride bool, generifyInterface bool) *ElementsStorage[T] {
+func NewElementsStorage[T comparable](allowOverride bool, reflectionOptions types.ReflectionOptions) *ElementsStorage[T] {
 	return &ElementsStorage[T]{
 		allowOverride:     allowOverride,
-		generifyInterface: generifyInterface,
+		generifyInterface: reflectionOptions,
 		namedElements:     map[string]genericAnyMap[T]{},
 		elements:          genericAnyMap[T]{},
 	}
 }
 
-func (s ElementsStorage[T]) ShouldGenerifyInterface() bool {
+func (s ElementsStorage[T]) ReflectOpts() types.ReflectionOptions {
 	return s.generifyInterface
 }
 
@@ -48,15 +51,21 @@ func (s *ElementsStorage[T]) SetNamed(elementType T, name string, value any) {
 	s.namedElements[name] = namedBinds
 }
 
-func (s ElementsStorage[T]) GetNamed(elementType T, name string) (any, bool) {
+func (s ElementsStorage[T]) GetNamed(elementType T, name string) (result any, err error) {
 	if elementMap, ok := s.namedElements[name]; ok && elementMap != nil {
-		result, subOk := elementMap[elementType]
-		return result, subOk
+		result, ok = elementMap[elementType]
+		if !ok {
+			err = utils.ErrElementNotRegistered
+		}
+		return
 	}
-	return nil, false
+	return nil, utils.ErrElementNotRegistered
 }
 
-func (s ElementsStorage[T]) Get(key T) (any, bool) {
-	result, ok := s.elements[key]
-	return result, ok
+func (s ElementsStorage[T]) Get(key T) (result any, err error) {
+	var ok bool
+	if result, ok = s.elements[key]; !ok {
+		err = utils.ErrElementNotRegistered
+	}
+	return
 }