+++ remy/internal/testelem/b/testtypes/types.go
@@ -1,0 +1,5 @@
+package testtypes
+
+type (
+	Syringe string
+)