+++ remy/internal/utils/generics.go
@@ -2,61 +2,36 @@ package utils
 
 import (
 	"fmt"
-	"github.com/wrapped-owls/goremy-di/remy/internal/types"
-	"reflect"
 	"strings"
 )
 
-func GetKey[T any](generifyInterface bool) types.BindKey {
-	elementType, isInterface := GetType[T]()
-	return TypeName(generifyInterface, elementType, isInterface)
-}
+// nilStr is the default const string representation of a nil type
+var nilStr = fmt.Sprintf("%T", nil)
 
-func GetElemKey(element any, generifyInterface bool) types.BindKey {
-	elementType, isInterface := GetElemType(element)
-	return TypeName(generifyInterface, elementType, isInterface)
-}
-
-// TypeName returns a string that defines the name of the given generic type.
-func TypeName(generifyInterface bool, elementType reflect.Type, isInterface bool) string {
-	if elementType == nil {
-		panic(ErrImpossibleIdentifyType)
-	}
-
-	if generifyInterface && isInterface {
-		var builder strings.Builder
-		builder.WriteString("interface { ")
-		for i := 0; i < elementType.NumMethod(); i++ {
-			if i > 0 {
-				builder.WriteString("; ")
-			}
-			builder.WriteString(
-				fmt.Sprintf(
-					"%s %s", elementType.Method(i).Name, elementType.Method(i).Type,
-				),
-			)
+func interfaceTypeName[T any](shouldGenerify bool, element T) (name string) {
+	if shouldGenerify {
+		if elementType, isInterface := GetElemType(element); isInterface {
+			return buildDuckInterfaceType(elementType.Elem())
 		}
-		builder.WriteString(" }")
-		return builder.String()
 	}
-	return fmt.Sprintf("%s/%s{###}%s", elementType.PkgPath(), elementType.Name(), fmt.Sprint(elementType))
-}
-
-func GetElemType(element any) (foundType reflect.Type, isInterface bool) {
-	foundType = reflect.TypeOf(element)
-	if foundType == nil {
-		panic(ErrImpossibleIdentifyType)
+	name = fmt.Sprintf("%T", element)
+	if name == nilStr {
+		name = fmt.Sprintf("%T", &element)
 	}
 	return
 }
 
-func GetType[T any]() (foundType reflect.Type, isInterface bool) {
-	var typeT T
-	foundType = reflect.TypeOf(typeT)
-	if foundType == nil {
-		// T is an interface
-		isInterface = true
-		foundType = reflect.TypeOf(&typeT)
+func TypeName[T any](shouldGenerify bool, elements ...T) (name string) {
+	var value T
+	if len(elements) > 0 {
+		value = elements[0]
+	}
+
+	name = fmt.Sprintf("%T", value)
+	if name == nilStr {
+		name = interfaceTypeName(shouldGenerify, value)
+	} else if strings.HasPrefix(name, "*") {
+		name = fmt.Sprintf("pointer_&%s", name)
 	}
 	return
 }